Fixed issue with exception being thrown when deleting config without data

+++ DataVisualization/DataVisualization.Core/ViewModels/DataBrowserViewModel.cs
@@ -52,7 +52,8 @@ namespace DataVisualization.Core.ViewModels
 
                 await Task.Run(() => {
                     _dataConfigurationService.DeleteConfigurationByName(selectedItem.DataName);
-                    _dataService.DeleteDataByName(selectedItem.DataName, progress);
+                    if(_dataService.Exists(selectedItem.DataName))
+                        _dataService.DeleteDataByName(selectedItem.DataName, progress);
                 });
 
                 _progressBarManager.CloseProgressBar();
